backend/requirements.txt
fastapi==0.63.0
uvicorn==0.16.0
supertokens-python==0.14.8
pysmb==1.2.9.1

